feature/컴포넌트 분리 작업
