+++ src/fb/myfirebase.js
@@ -1,6 +1,6 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-import { getAuth } from "firebase/auth";
+import { getAuth, onAuthStateChanged } from "firebase/auth";
 import { getDatabase } from "firebase/database";
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -20,3 +20,16 @@ export const app = initializeApp(firebaseConfig);
 
 export const AUTH = getAuth(app);
 export const DB = getDatabase(app);
+
+export const currentUser = onAuthStateChanged(AUTH, (user) => {
+  if (user) {
+    console.log(user);
+    const uid = user.uid;
+    console.log(uid);
+    return user;
+    // ...
+  } else {
+    // User is signed out
+    // ...
+  }
+});
