+++ src/redux/config/configStore.js
@@ -1,10 +1,11 @@
 import { createStore } from "redux";
 import { combineReducers } from "redux";
+import authState from "redux/modules/authState";
 import modalState from "redux/modules/modalState";
 const rootReducer = combineReducers({
-  modalState
+  modalState,
+  authState
 });
-
 const store = createStore(rootReducer);
 
 export default store;
